Improve data reliability
